+++ src/ankh/helm/helm.go
@@ -269,159 +269,11 @@ func Template(ctx *ankh.ExecutionContext, ankhFile ankh.AnkhFile) (string, error
 			}
 			finalOutput += chartOutput
 		}
+	} else {
+		ctx.Logger.Infof(
+			"%s does not contain any charts. Template commands only operate on ankh.yaml files containing charts",
+			ctx.AnkhFilePath)
 	}
 
 	return finalOutput, nil
-}
-
-func getChartFileContent(ctx *ankh.ExecutionContext, path string, useContext bool, key string) ([]byte, error) {
-	var result []byte
-	bytes, err := ioutil.ReadFile(fmt.Sprintf("%s", path))
-	if err == nil {
-
-		if useContext {
-			bytes, err = CreateReducedYAMLFile(path, key)
-			if err != nil {
-				return result, err
-			}
-		}
-
-		result = bytes
-	} else {
-		ctx.Logger.Debugf("%s not found", path)
-	}
-
-	if len(bytes) > 0 {
-		result = append([]byte("---\n# Source: "+path+"\n"), bytes...)
-	}
-
-	return result, nil
-}
-
-func InspectValues(ctx *ankh.ExecutionContext, chart ankh.Chart, ankhFile ankh.AnkhFile) (string, error) {
-	var result string
-
-	ctx.Logger.Debug("Inspecting values for chart %s", chart.Name)
-
-	result += "---\n# Chart: " + chart.Name
-	result += fmt.Sprintf("\n# Source: %s\n", ctx.AnkhFilePath)
-
-	type Values struct {
-		DefaultValues    map[string]interface{} `yaml:"default_values"`
-		Values           interface{}
-		ResourceProfiles interface{} `yaml:"resource_profiles"`
-	}
-
-	values := Values{}
-	if ctx.UseContext {
-		values = Values{
-			DefaultValues:    chart.DefaultValues,
-			Values:           chart.Values[ctx.AnkhConfig.CurrentContext.Environment],
-			ResourceProfiles: chart.ResourceProfiles[ctx.AnkhConfig.CurrentContext.Profile],
-		}
-	} else {
-		values = Values{
-			DefaultValues:    chart.DefaultValues,
-			Values:           chart.Values,
-			ResourceProfiles: chart.ResourceProfiles,
-		}
-	}
-
-	out, err := yaml.Marshal(values)
-	if err != nil {
-		return "", nil
-	}
-	result += string(out)
-
-	files, err := FindChartFiles(ctx, chart.Name, chart.Version, ankhFile)
-	if err != nil {
-		return "", nil
-	}
-
-	bytes, err := getChartFileContent(ctx, files.AnkhResourceProfilesPath, ctx.UseContext, ctx.AnkhConfig.CurrentContext.Profile)
-	if err != nil {
-		return "", nil
-	}
-	if len(bytes) > 0 {
-		result += string(bytes)
-	}
-
-	bytes, err = getChartFileContent(ctx, files.AnkhValuesPath, ctx.UseContext, ctx.AnkhConfig.CurrentContext.Environment)
-	if err != nil {
-		return "", nil
-	}
-	if len(bytes) > 0 {
-		result += string(bytes)
-	}
-
-	bytes, err = getChartFileContent(ctx, files.ValuesPath, false, "")
-	if err != nil {
-		return "", nil
-	}
-	if len(bytes) > 0 {
-		result += string(bytes)
-	}
-
-	return result, nil
-}
-
-func InspectChart(ctx *ankh.ExecutionContext, chart ankh.Chart, ankhFile ankh.AnkhFile) (string, error) {
-	var result string
-
-	ctx.Logger.Debug("Inspecting chart.yaml for chart %s", chart.Name)
-
-	currentContext := ctx.AnkhConfig.CurrentContext
-	result += fmt.Sprintf("# Chart: %s\n", chart.Name)
-	files, err := FindChartFiles(ctx, chart.Name, chart.Version, ankhFile)
-	if err != nil {
-		return "", err
-	}
-	helmArgs := []string{"helm", "inspect", "chart", "--kube-context", currentContext.KubeContext}
-
-	helmArgs = append(helmArgs, files.ChartDir)
-
-	ctx.Logger.Debugf("running helm command %s", strings.Join(helmArgs, " "))
-
-	helmCmd := exec.Command(helmArgs[0], helmArgs[1:]...)
-	helmOutput, err := helmCmd.CombinedOutput()
-
-	if err != nil {
-		return "", fmt.Errorf("error running the helm command: %s", helmOutput)
-	}
-
-	result += string(helmOutput)
-
-	return result, nil
-}
-
-func InspectTemplates(ctx *ankh.ExecutionContext, chart ankh.Chart, ankhFile ankh.AnkhFile) (string, error) {
-	var result string
-
-	ctx.Logger.Debug("Inspecting templates for chart %s", chart.Name)
-	files, err := FindChartFiles(ctx, chart.Name, chart.Version, ankhFile)
-	if err != nil {
-		return "", err
-	}
-
-	dir := files.ChartDir + "/templates/"
-
-	var templates []os.FileInfo
-	templates, err = ioutil.ReadDir(dir)
-	if err != nil {
-		return "", err
-	}
-
-	result += "---\n# Chart: " + chart.Name
-	for _, template := range templates {
-		result += fmt.Sprintf("\n# Source: %s/templates/%s\n", chart.Name, template.Name())
-		path := dir + template.Name()
-		bytes, err := ioutil.ReadFile(path)
-		if err != nil {
-			return "", err
-		}
-
-		result += string(bytes)
-	}
-
-	return result, nil
 }
